requirements.txt
flask
scikit-learn
numpy
pandas
gunicorn 
python-dotenv
google-generativeai